+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityControls.java
@@ -215,7 +215,7 @@ final class EntityControls {
 
 		@Override
 		public void accept(Entity inserted) {
-			comboBoxModel.add(inserted);
+			comboBoxModel.addItem(inserted);
 			comboBoxModel.setSelectedItem(inserted);
 		}
 	}